Remove non-free or unidentified files.
This effort is to bring the project in compliance with most
distributions guidelines such as the GNU Free Software Distribution
Guidelines (GNU FSDG) [0]: the files without a clear license or a
non-free license are removed.

This is a work in progress; some files such as README.md and other
project files can be easily re-added when the project owners introduce
a LICENSE file with some explanatory text.

This is a rather strict purge; I haven't tried contacting the authors
or going through hoops to recover it from the original place the files
were found; I've simply assumed that if there's no license information
within the files, then they aren't free.  This is relaxed for
non-functional sources as the .slangp configuration or images, as
non-functional data does not require to be freely licensed, per the
GNU FSDG.

[0]  https://www.gnu.org/distros/free-system-distribution-guidelines.html

Fixes: https://github.com/libretro/slang-shaders/issues/150